Task-Management-App
This is the Task Management App which has a User panel and an Admin panel. Users have access to the user panel and Admin has Access to the admin panel to add or delete the detail or to made the amendement. The accessibility key for the admin panel are given below:
 adminName: "Arshman",
 password: "Arsh#786",
 email: "Arsh@786"
Copy these things to access the admin panel.